Moved TaxTimer into Townservice as requested.
